Tring to integrate classes
