+++ Layout/MainLayout.razor
@@ -1,5 +1,8 @@
 ﻿@inherits LayoutComponentBase
 @using MyBlazorPwa.Layout
+@using Microsoft.AspNetCore.Components
+
+
 
 <div class="page">
     <div class="sidebar">
