--- MyBlazorPwa.generated.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PMListener", "HTTP_Listener_Worker.NET\PMListener.csproj", "{D7CAD4F5-16FC-4E5F-982E-9B4A9389F06C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyBlazorPwa", "MyBlazorPwa\MyBlazorPwa.csproj", "{FA71EA58-9682-4C98-9A9C-A0E5613DF0E3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D7CAD4F5-16FC-4E5F-982E-9B4A9389F06C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D7CAD4F5-16FC-4E5F-982E-9B4A9389F06C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D7CAD4F5-16FC-4E5F-982E-9B4A9389F06C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D7CAD4F5-16FC-4E5F-982E-9B4A9389F06C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FA71EA58-9682-4C98-9A9C-A0E5613DF0E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FA71EA58-9682-4C98-9A9C-A0E5613DF0E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FA71EA58-9682-4C98-9A9C-A0E5613DF0E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FA71EA58-9682-4C98-9A9C-A0E5613DF0E3}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6F26438B-E51E-4C4E-8745-AFF53E547362}
-	EndGlobalSection
-EndGlobal
