+++ Pages/ProtocolVisualizer.razor
@@ -1,0 +1,76 @@
+
+/* Responsible for displaying protocol data on the screen according to user settings and preferences. */
+
+@code {
+    [Parameter]
+    public List<ProtocolData> Protocols { get; set; } // Lista de protocolos a serem exibidos
+
+    private bool isExpanded; // Para controlar a expansão do visualizador
+
+    private void ToggleExpand()
+    {
+        isExpanded = !isExpanded; // Alterna a expansão
+    }
+}
+
+<div class="protocol-visualizer">
+    @foreach (var protocol in Protocols)
+    {
+        <div class="protocol-card" @onclick="ToggleExpand">
+            <h3>Protocolo: @protocol.Numero</h3>
+            @if (isExpanded)
+            {
+                <div class="protocol-visualizer">
+    @foreach (var protocol in Protocols)
+    {
+        <div class="protocol-card" @onclick="ToggleExpand">
+            <h3>Protocolo: @protocol.Header.Numero/@protocol.Header.Ano</h3>
+            @if (isExpanded)
+            {
+                <div>
+                    <p><strong>Interessado:</strong> @protocol.Header.Interessado</p>
+                    <p><strong>Solicitante:</strong> @protocol.Header.Solicitante</p>
+                    <p><strong>Inspetoria:</strong> @protocol.Header.Inspetoria</p>
+                    <p><strong>Assunto:</strong> @protocol.Header.Assunto</p>
+                    <p><strong>Origem:</strong> @protocol.Header.Origem</p>
+                    <p><strong>Situação:</strong> @protocol.Header.Situacao</p>
+                    <p><strong>Destino:</strong> @protocol.Header.Destino</p>
+                    <p><strong>Sigiloso:</strong> @(protocol.Header.Sigiloso == true ? "Sim" : "Não")</p>
+                    <p><strong>Data de Emissão:</strong> @protocol.Header.DataEmissao?.ToString("dd/MM/yyyy")</p>
+                    <p><strong>Descrição:</strong> @protocol.Header.Descricao</p>
+                    
+                    <h4>Movimentações:</h4>
+                    @foreach (var move in protocol.Moves)
+                    {
+                        <div class="move-card">
+                            <p><strong>Passo:</strong> @move.Passo</p>
+                            <p><strong>Usuário de Origem:</strong> @move.UsuarioOrigem</p>
+                            <p><strong>Usuário de Destino:</strong> @move.UsuarioDestino</p>
+                            <p><strong>Setor de Origem:</strong> @move.SetorOrigem</p>
+                            <p><strong>Setor de Destino:</strong> @move.SetorDestino</p>
+                            <p><strong>Descrição:</strong> @move.Descricao</p>
+                            <p><strong>Data:</strong> @move.Data?.ToString("dd/MM/yyyy")</p>
+                            <p><strong>Hora:</strong> @move.Hora?.ToString(@"hh\:mm\:ss")</p>
+                            <p><strong>Sigiloso:</strong> @(move.Sigiloso == true ? "Sim" : "Não")</p>
+                        </div>
+                    }
+                </div>
+            }
+        </div>
+    }
+</div>
+            }
+        </div>
+    }
+</div>
+
+<style>
+    .protocol-card {
+        border: 1px solid #ccc;
+        margin: 10px;
+        padding: 10px;
+        cursor: pointer;
+    }
+</style>
+
+}
