--- Pages/WebSocket.razor
@@ -1,100 +1,0 @@
-
-
-@page "/websocket"
-@using System.Threading.Tasks
-@using PM.classes
-@inject WebSocketService WebSocketService
-
-<h3>WebSocket Example</h3>
-<button @onclick="ConnectWebSocket">Connect</button>
-
-<form id="textForm" @onsubmit="SendMessage">
-  <label for="userInput">Insira seu texto:</label>
-  <input type="text" id="userInput" @bind="userInput" name="userInput">
-  <button type="submit">Enviar</button>
-</form>
-
-@page "/"
-
-
-<PageTitle>Protocolos</PageTitle>
-
-@code {
-    private List<ProtocolData> protocols;
-
-    protected override async Task OnInitializedAsync()
-    {
-        protocols = await LoadProtocols(); // Carrega protocolos de algum lugar, como um serviço ou arquivo
-    }
-
-    private async Task<List<ProtocolData>> LoadProtocols()
-    {
-        // Lógica para carregar protocolos (por exemplo, de um arquivo JSON ou de um banco de dados)
-        return new List<ProtocolData>(); // Retornar uma lista de ProtocolData preenchida
-    }
-}
-
-<h1>Visualização de Protocolos</h1>
-
-<ProtocolVisualizer Protocols="protocols" />
-
-
-@if (isLoading)
-{
-    <p>Carregando...</p>
-}
-else if (receivedData != null)
-{
-    <h4>Dados Recebidos:</h4>
-    <pre>@receivedData</pre>
-}
-else if (!string.IsNullOrEmpty(receivedMessage))
-{
-    <p>@receivedMessage</p>
-}
-
-@code {
-    private string receivedMessage;
-    private string userInput;
-    private bool isLoading = false;
-    private string receivedData;
-
-    private async Task ConnectWebSocket()
-    {
-        await WebSocketService.ConnectAsync();
-    }
-
-    private async Task SendMessage()
-    {
-        isLoading = true;
-        receivedData = null;
-        receivedMessage = null;
-        StateHasChanged();
-
-        await WebSocketService.SendAsync(userInput);
-        await ReceiveMessage();
-
-        isLoading = false;
-        StateHasChanged();
-    }
-
-    private async Task ReceiveMessage()
-    {
-        receivedMessage = await WebSocketService.ReceiveAsync();
-
-        // Parse JSON if needed
-        if (!string.IsNullOrEmpty(receivedMessage))
-        {
-            try
-            {
-                receivedData = receivedMessage;
-            }
-            catch
-            {
-                receivedData = "Erro ao processar os dados recebidos.";
-            }
-        }
-    }
-}
-
-}
