+++ _Imports.razor
@@ -9,4 +9,4 @@
 @using Microsoft.JSInterop
 @using MyBlazorPwa
 @using MyBlazorPwa.Layout
-@using PM.classes
+
