--- classes/ProtocolLoader.cs
@@ -1,3 +1,0 @@
-namespace PM.classes{
-
-}
